gcloud_function/functions/requirements.txt
firebase_admin
firebase_functions==0.1.1
google-auth
google-cloud